Record visualizer consolidation in steps summary
